chore: split shared list table concerns
