+++ src/js/components/common/KebabMenu.tsx
@@ -113,6 +113,52 @@ const usePopoverPlacement = ({ isOpen, triggerRef, popoverRef }: PopoverBehaviou
 	return isFlipped
 }
 
+const handlePopoverKeyDown = (
+	event: KeyboardEvent,
+	{ closeMenu, popoverRef }: Pick<PopoverBehaviourOptions, 'closeMenu' | 'popoverRef'>
+) => {
+	if ('Escape' === event.key) {
+		event.preventDefault()
+		closeMenu()
+		return
+	}
+
+	if (!popoverRef.current) {
+		return
+	}
+
+	const focusable = Array.from(popoverRef.current.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR))
+	const active = document.activeElement
+	const currentIndex = active instanceof HTMLElement ? focusable.indexOf(active) : -1
+
+	if ('ArrowDown' === event.key || 'ArrowUp' === event.key || 'Home' === event.key || 'End' === event.key) {
+		if (0 < focusable.length) {
+			event.preventDefault()
+			const offset = 'ArrowDown' === event.key ? currentIndex + 1 : currentIndex - 1
+			const target = 'Home' === event.key
+				? 0
+				: 'End' === event.key ? focusable.length - 1 : (offset + focusable.length) % focusable.length
+			focusable[target].focus()
+		}
+
+		return
+	}
+
+	if ('Tab' !== event.key) {
+		return
+	}
+
+	if (0 === focusable.length) {
+		event.preventDefault()
+	} else if (event.shiftKey && 0 >= currentIndex) {
+		event.preventDefault()
+		focusable[focusable.length - 1].focus()
+	} else if (!event.shiftKey && (-1 === currentIndex || currentIndex === focusable.length - 1)) {
+		event.preventDefault()
+		focusable[0].focus()
+	}
+}
+
 // Close the open popover on outside clicks or Escape, and trap Tab focus
 // within it while it remains open.
 const usePopoverDismissal = ({ isOpen, setIsOpen, closeMenu, containerRef, popoverRef }: PopoverBehaviourOptions) => {
@@ -127,48 +173,8 @@ const usePopoverDismissal = ({ isOpen, setIsOpen, closeMenu, containerRef, popov
 			}
 		}
 
-		const handleKeyDown = (event: KeyboardEvent) => {
-			if ('Escape' === event.key) {
-				event.preventDefault()
-				closeMenu()
-				return
-			}
-
-			if (!popoverRef.current) {
-				return
-			}
-
-			const focusable = Array.from(popoverRef.current.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR))
-			const active = document.activeElement
-			const currentIndex = active instanceof HTMLElement ? focusable.indexOf(active) : -1
-
-			if ('ArrowDown' === event.key || 'ArrowUp' === event.key || 'Home' === event.key || 'End' === event.key) {
-				if (0 < focusable.length) {
-					event.preventDefault()
-					const offset = 'ArrowDown' === event.key ? currentIndex + 1 : currentIndex - 1
-					const target = 'Home' === event.key
-						? 0
-						: 'End' === event.key ? focusable.length - 1 : (offset + focusable.length) % focusable.length
-					focusable[target].focus()
-				}
-
-				return
-			}
-
-			if ('Tab' !== event.key) {
-				return
-			}
-
-			if (0 === focusable.length) {
-				event.preventDefault()
-			} else if (event.shiftKey && 0 >= currentIndex) {
-				event.preventDefault()
-				focusable[focusable.length - 1].focus()
-			} else if (!event.shiftKey && (-1 === currentIndex || currentIndex === focusable.length - 1)) {
-				event.preventDefault()
-				focusable[0].focus()
-			}
-		}
+		const handleKeyDown = (event: KeyboardEvent) =>
+			handlePopoverKeyDown(event, { closeMenu, popoverRef })
 
 		document.addEventListener('mousedown', handlePointerDown)
 		document.addEventListener('keydown', handleKeyDown)
